Update Port Number & Set up Instruction (#8)
What does this Pull Request accomplish?
This pull request updates the Port number of the server, user manual and set up instruction.

Why should this Pull Request be merged?
The port number that the server used is not recommended and so it is updated. The user manual and set up instruction is updated accordingly.

What testing has been done?
Testing is not required.